Initial work for supporting HA

+++ app/Controller/TasksController.php
@@ -56,6 +56,9 @@ class TasksController extends AppController
         if (!$this->_isSiteAdmin()) {
             throw new MethodNotAllowedException('You are not authorised to do that.');
         }
+        if (!$this->_isHAPrimary()) {
+           throw new MethodNotAllowedException('The is not the primary instance in a HA deployment.');
+        }
 
         $task = $this->Task->find('first', array(
             'recursive' => -1,
@@ -86,6 +89,9 @@ class TasksController extends AppController
         if (!$this->_isSiteAdmin()) {
             throw new MethodNotAllowedException('You are not authorised to do that.');
         }
+        if (!$this->_isHAPrimary()) {
+           throw new MethodNotAllowedException('The is not the primary instance in a HA deployment.');
+        }
 
         $this->set('dropdownData', $this->__getDropdownData());
 
@@ -110,6 +116,9 @@ class TasksController extends AppController
     {
         if (!$this->_isSiteAdmin()) {
             throw new MethodNotAllowedException('You are not authorised to do that.');
+        }
+        if (!$this->_isHAPrimary()) {
+           throw new MethodNotAllowedException('The is not the primary instance in a HA deployment.');
         }
 
         $this->set('dropdownData', $this->__getDropdownData());
@@ -200,6 +209,10 @@ class TasksController extends AppController
         if (!$this->_isSiteAdmin()) {
             throw new MethodNotAllowedException('You are not authorised to do that.');
         }
+        if (!$this->_isHAPrimary()) {
+           throw new MethodNotAllowedException('The is not the primary instance in a HA deployment.');
+        }
+
         $this->CRUD->delete($id);
         if ($this->IndexFilter->isRest()) {
             return $this->restResponsePayload;
@@ -211,7 +224,9 @@ class TasksController extends AppController
         if (!$this->_isSiteAdmin()) {
             throw new MethodNotAllowedException('You are not authorised to do that.');
         }
-
+        if (!$this->_isHAPrimary()) {
+           throw new MethodNotAllowedException('The is not the primary instance in a HA deployment.');
+        }
         $task = $this->Task->find('first', array(
             'recursive' => -1,
             'conditions' => array('Task.id' => $id)
